Signal no longer needs exchange of phone numbers
This has been the case since February 2024. Reference:
 - https://signal.org/blog/phone-number-privacy-usernames/
 - https://support.signal.org/hc/en-us/articles/6712070553754-Phone-Number-Privacy-and-Usernames

The .xlsx file format is binary. To update the spreadsheet, I opened it in
Libreoffice Calc 4:7.4.7-1+deb12u1, changed cell Y23 to "No", and saved it
with the current extension using all default options. I do not know if there
is a way to make changes in this format directly legible to git, but it is
possible to compare by exporting the spreadsheet before and after to CSV, then
running a standard text diff on the result.